server/requirements.txt
absl-py==0.11.0
aniso8601==8.1.0
appdirs==1.4.4
autopep8==1.5.4
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-JWT-Extended==3.25.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
livereload==2.6.3
lunr==0.5.8
Markdown==3.3.3
MarkupSafe==1.1.1
mkdocs==1.1.2
mkdocs-material==6.2.7
mkdocs-material-extensions==1.0.1
nltk==3.5
ortools==8.1.8487
pipenv==2020.11.15
protobuf==3.14.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pymdown-extensions==8.1.1
pyodbc==4.0.30
pypyodbc==1.3.4.3
python-decouple==3.4
python-dotenv==0.15.0
python-i18n==0.3.9
pytz==2020.5
PyYAML==5.4.1
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
tornado==6.1
tqdm==4.56.0
virtualenv==20.4.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
